Extract strings; unify icon + text change function calls in download bar view

+++ app/src/main/java/org/jnanaprabodhini/happyteacher/extension/NumberExtensions.kt
@@ -4,6 +4,6 @@ package org.jnanaprabodhini.happyteacher.extension
  * Created by grahamearley on 9/29/17.
  */
 
-fun Long.toMegabyteFromByte(): Long {
-    return  this / 1000000
+fun Long.toMegabyteFromByte(): Double {
+    return  this / 1000000.toDouble()
 }
